t: add existing/missing/bad object SSH push tests
Our t/t-push-failures-local.sh test script includes a number of tests
which validate the behaviour of the "git push" command when Git LFS
objects are missing or corrupt.

All of these tests exercise the HTTP-based Git LFS object transfer
protocol, as they predate the development and implementation of the
"pure" SSH-based Git LFS object transfer protocol in PR #4446
in 2021.

In a subsequent commit in this PR we expect to update the Git LFS
client to remove some non-functional code which attempts to recreate
missing Git LFS objects during push operations.  In many cases this
change will cause the client to report missing objects in an earlier
stage of push operations than it does now.  We also expect to reword
the error message the client will output in such cases.

Before we make these changes, we first introduce additional versions of
all the tests in the t/t-push-failures-local.sh script, adjusting them
to exercise the SSH-based Git LFS object transfer protocol rather than
the HTTP-based one.  We also introduce a new refute_remote_object()
test helper function, akin to the existing refute_server_object()
helper function but like the assert_remote_object() function able
to operate without the use of the HTTP-based Git LFS object transfer
protocol.

Among our new tests, one test in particular, the "push reject missing
object (lfs.allowincompletepush false) (git-lfs-transfer)" test,
validates some logic in the Batch() method of the SSHBatchClient
structure in our "tq" package that we expect to refactor in a
subsequent commit in this PR.

At present, this Batch() method operates in the same manner as the
Batch() method of the tqClient, which implements the HTTP-based
object transfer protocol.  Both methods start by iterating over the
set of objects to be transferred and copying the Missing fields from
the per-object Transfer structures into a local map.  After performing
the actual Batch API request (either over HTTP or SSH), the both
methods then iterate over the set of object records returned by the
server, which are also represented as Transfer structures, and set
these structures' Missing flags based on the values saved in the
methods' local maps.

This behaviour ensures that when the enqueueAndCollectRetriesFor()
method of the TransferQueue structure in our "tq" package iterates
over the per-object Transfer structures representing the server's
response, if the server has requested an upload of the object (meaning
the server lacks a copy of the object), and if the object's Missing
flag is set (meaning the client lacks a local copy of the object data),
then the push operation can be abandoned immediately, instead of
proceeding further and only reporting the missing object later
when attempting to upload the object's data.

As noted above, we expect to refactor this logic in later commits in
this PR, specifically by removing the use of local maps in the Batch()
methods while retaining the ability of the client to abandon push
operations at the batch request stage as soon as the client determines
that an object is missing both locally and on the remote server.

The new tests we add in this commit will therefore provide greater
confidence that these subsequent changes will not cause regressions in
the existing behaviour of the client when performing SSH-based push
operations.