t,tq: log trace when missing object found in batch
At present, during push operations, the Git LFS client outputs different
error messages when an object file is missing depending on whether or
not a file exists in the working tree at the path associated with the
Git LFS pointer corresponding to the object.  The file in the working
tree may have any content, or be entirely empty; its simple presence
is enough to change the error message output by the Git LFS client and
the point at which the client abandons the upload push operation, so
long as the "lfs.allowIncompletePush" configuration option is set to
its default value of "false".

Under these conditions, the client also abandons the push operation
as soon as it determines that the object is missing on the remote
server as well as on the local system.  Under other conditions, such
as when a file (with any content) exists in the working tree at the
path associated with the missing object, the client continues the
push operation until it attempts to upload the object's data, at
which point it skips the object but continues to try to upload any
other objects in its queue.

In a subsequent commit in this PR we will remove the code which
causes this variable behaviour, which will simplify our handling of
missing objects during upload transfer operations, and cause the
client to output a consistent error message in all cases where an
object file is missing and the "lfs.allowIncompletePush" configuration
option is set to "false".  Our changes will also cause the client to
abandon the push operation as soon as it determines an object to be
missing both locally and on the remote server.

Our t/t-pre-push.sh and t/t-push-failures-local.sh test scripts include
a number of tests which validate the behaviour of the "git lfs pre-push"
and "git push" commands when Git LFS objects are missing or corrupt.

Several of these tests simulate the specific conditions described
above, and check that the push operation has been abandoned immediately
after a Batch API request by checking for the presence of a specific
error message.  This message was originally introduced into the
ensureFile() method of the uploadContext structure in our "commands"
package in commit fea77e1c6f715964d6ebbd8e71374c1896cad9ca of PR #3398,
but since commit 1412d6e47af351d678f50cf9a5d6e0bc9c92c919 of PR #3634
was been output by the enqueueAndCollectRetriesFor() method of the
TransferQueue structure in our "tq" package.

When we revise this message to be consistent with the one output
by the ReportErrors() method of the uploadContext structure, which
we expect to do in a subsequent commit in this PR, our tests will
no longer be able to check for the unique message output by the
enqueueAndCollectRetriesFor() method to confirm that the push
operation has been abandoned immediately after receiving a response
from the remote server.

As we would like our tests to continue to validate this behaviour by
checking for the presence of a specific error message output by the
enqueueAndCollectRetriesFor() method, we update that method now to
output a trace log message, and then revise the relevant tests to
also set the GIT_TRACE environment variable and then check for this
message in the logs generated by the "git lfs pre-push" or "git push"
commands.