docs/man/git-lfs-config.adoc: fix missing term
When we converted our manual page source files from the Ronn format to
the AsciiDoc format in commit 0c66dcf15d988802255c3133cd5ab8105cbf0ef7
of PR #5054, a few terms were accidentally dropped if they were delimited
with angle brackets but not also enclosed with backtick characters.

In commit 219518ec1fe136c89930d5e25ea0291e749be875 of PR #5414 we
restored one such missing term in the git-lfs-logs(1) manual page,
and in a prior commit in this PR we restored another in the
git-lfs-checkout(1) manual page.

We now restore another missing term, this time in the git-lfs-config(5)
manual page, in the entry for the "lfs.customtransfer.<name>.path"
configuration option, where the "<name>" term was accidentally removed
from a sentence in which it was the first word.

We also take the opportunity to add a leading definite article to the
sentence so to make it slightly more idiomatic and readable.