Fix migrations where `.git/objects` is a symlink
Android's `repo` tool creates `.git/objects` as a symlink to a place
outside the `.git` directory for unknown reasons.  When that happens,
Git operations where we set the working directory to that directory
fail.  This is because when Git performs discovery on the working
directory, it first canonicalizes the path, which takes it outside of
the `.git` directory and it then fails to find anything that looks like
a Git repository.

In our codebase, we set the path to the object directory root in a
couple of places which cause this problem.  To avoid this problem,
simply join the repository root with `..`, which is always applied
before canonicalization and will result in the path being properly set
to the `.git` directory, which results in `git lfs migrate import`
working correctly in this configuration.