docs/man/git-lfs-checkout.adoc: fix missing term
When we converted our manual page source files from the Ronn format to
the AsciiDoc format in commit 0c66dcf15d988802255c3133cd5ab8105cbf0ef7
of PR #5054, a few terms were accidentally dropped if they were delimited
with angle brackets but not also enclosed with backtick characters.

In commit 219518ec1fe136c89930d5e25ea0291e749be875 of PR #5414 we
restored one such missing term in the git-lfs-logs(1) manual page.

We now restore another missing term, this time in the description section
of the git-lfs-checkout(1) manual page, where the "<glob-pattern>"
term was accidentally removed.

We also take the opportunity to rephrase the sentence containing this
term slightly so as to avoid the use of a trailing plural "s" character,
as that causes formatting errors when rendering the AsciiDoc source into
HTML or roff.