docs/man/git-lfs-migrate.adoc: amend units sublist
In our git-lfs-migrate(1) manual page, one of the entries in the list
of options supported by the "git lfs migrate info" subcommand describes
the --unit option, and enumerates the set of storage capacity units that
we accept for that option.  In this option's entry we provide two lists,
one of decimal SI units (e.g., kb, MB) and one of binary IEC units
(e.g., KiB, MiB), with the intent that they should each form a single
entry in a second-level list nested within the top-level list of command
options.

However, when we converted our manual page source files from the Ronn
format to AsciiDoc in commit 0c66dcf15d988802255c3133cd5ab8105cbf0ef7
of PR #5054, this second-level list was flattened into continuous text,
so we restore it now.

In this commit we simply match the formatting already in use in several
second-level lists in the git-lfs-config(5) manual page, although it
does not render properly in any of our output formats.  We anticipate
that in a subsequent PR we will adjust the formatting of all of these
second-level lists so they render correctly when converted to HTML and
roff by Asciidoctor.