docs/man/git-lfs-dedup.adoc: add test option
Since the "git lfs dedup" command was introduced in PR #3753 it has
accepted a --test option, which causes the command to check whether the
current operating system and file system support the copy-on-write
functionality the command uses to create clones or reflinks of Git LFS
files.

However, this option is not documented in our git-lfs-dedup(1) manual
page, so we now add an entry for it into the list of options supported
by the command.