commands,docs/man: make JSON options consistent
A number of our commands accept a --json option, which causes the
commands to write the details of their operation in JSON format to
standard output, unless the command fails for some reason.

Our manual pages document this option for the majority of commands that
support it, but not for the "git lfs track" command, which was updated
in commit 6f37e98ae8279bb1db1ac6b77303ce1df442d863 of PR #5006 to also
provide JSON output when passed the --json option.

As well, the "git lfs track" command does not yet accept the same
short -j version of the --json option as all the other commands.

We address these issues by adding support for the -j short option
to the "git lfs track" command and documenting both the --json and -j
options in the corresponding manual page.

Further, we take the opportunity to update the descriptions of the
same options in our git-lfs-fetch(1), git-lfs-ls-files(1), and
git-lfs-status(1) manual pages so they use the same language found in
the manual pages for the other Git LFS commands that provide the option
of JSON-formatted output, and describe the how the --json option
interoperates with other options.