docs/man/git-lfs-prune.adoc: amend configs list
In our git-lfs-prune(1) manual page we dedicate one section to a
explanation of how the Git LFS client determines whether an object
is sufficiently "recent" to be retained and not pruned.

In this section we list the Git configuration options that control
the expiry and retention periods for objects, including settings
such as "lfs.fetchRecentCommitsDays", "lfs.fetchRecentRefsDays", and
"lfs.pruneOffsetDays".  The list contains two entries, one for
the "lfs.pruneOffsetDays" option and one for three of the
"lfs.fetchRecent*" options.

When the git-lfs-prune(1) manual page was first added in commit
bd72983c05d32115e24d268eccb1c2a24907c26a of PR #742, this list
of the Git configuration options was formatted such that each entry
began with the option name or names as headers on separate lines,
followed by a description of their effects as a separate paragraph.

When we converted our manual pages from the Ronn format to AsciiDoc
in commit 0c66dcf15d988802255c3133cd5ab8105cbf0ef7 of PR #5054, the
list was formatted as an unordered list where the option name or names
and the subsequent description were simply all placed together as a
single paragraph.

To help make this section of the git-lfs-prune(1) manual page slightly
more readable, we adjust its formatting now to use an AsciiDoc
description list rather than an unordered list.  By doing so, we can
restore the option names as distinct headers, each of which will appear
on a separate line when we render the source file into HTML or roff
with Asciidoctor.  The description text of each entry will also then be
rendered as an indented paragraph following the entry's headers.