t: test checkout --to option with external paths
In commit cf7f9678b3d2929425d0671c099c2ef5621c0929 of PR #3296 we
introduced support for the --to, --ours, --theirs, and --base options in
the "git lfs checkout" command, and added a "checkout: conflicts" test
to our t/t-checkout.sh test script to validate the behaviour of these
new options.

This test checks that when the --to option is provided along with one
of the other options, the appropriate patch diff output is written to
the file specified with the --to option.

However, at present, we only perform these checks using local file
names, although our git-lfs-checkout(1) manual page states that a file
external to the working tree may be specified with the --to option.

We therefore revise our test to ensure that we run the "git lfs checkout"
command with --to option arguments specifying files outside of the
working tree, in one case using a relative path and in two other cases
an absolute path.  With the absolute path check we also confirm that
the command will create any directories in the path that do not exist,
as well as traverse any symbolic links to directories so long as the
directories exist.  (Note that if the filename component of the path
is a link to a directory, an error will occur when the Git LFS client
attempts to open it for writing, so we do not test this case.)

We also perform these checks again after changing the current working
directory to a subdirectory of the work tree, this time using relative
paths with ".." path components to specify the file in the repository
for which a patch diff should be generated.  By performing these checks
we verify that the "git lfs checkout" command supports relative paths
from a current working directory which is not the root of the work tree.
In a subsequent commit we will update the "git lfs checkout" command
so that it changes the current working directory before generating any
patch diff output, at which time these additional checks will help
demonstrate that our changes still support the use of paths relative
to the working directory in which the user originally runs the command.

On Windows, true symbolic link support is not enabled by default and
not supported on all filesystems or by all versions of Windows.  We
therefore only test the "git lfs checkout" command with a path for
the --to option which traverses a symbolic link if we can determine
that symbolic links can actually be created on the current Windows
system.  To do this we introdce a new has_native_symlinks() test helper
function, which returns a successful exit code only if the current
system supports the creation of symbolic link.  We expect to make
additional use of this helper function in subsequent commits.

On Unix systems, our has_native_symlinks() always returns a successful
(i.e., zero) exit code.  On Windows it first tries to enable native
symbolic link support in the Cygwin or MSYS2 environments, and then
returns a successful exit code only if a test symbolic link is actually
created by the ln(2) command.  This Unix command is emulated in the
MSYS2 and Cygwin environments, which are in turn used by the Git Bash
environment in which we run our test suite on Windows.  To check whether
a true Windows symbolic link has been created, we check the results of
a query made with the Windows "fsutil reparsepoint" command.  See,
for reference:

  https://cygwin.com/cygwin-ug-net/using.html#pathnames-symlinks
  https://www.msys2.org/docs/symlinks/
  https://learn.microsoft.com/en-us/windows/apps/get-started/enable-your-device-for-development
  https://learn.microsoft.com/en-us/windows-server/administration/windows-commands/fsutil-reparsepoint

Fortunately, the GitHub Actions Windows runners we use to run our
CI test suite have Developer Mode enabled, and so true symbolic links
may be created on these systems.

Finally, we adjust the order in which we check the contents of the
files output by the "git lfs checkout" commands so as to match the
order in which we run those commands.