t/t-{checkout,pull}.sh: test changed files skipped
In PR #527 we introduced the "git lfs checkout" and "git lfs pull"
commands and added some initial tests of those commands to our
test suite, which we have subsequently expanded over time.

Since that PR, these commands have consistently used the
DecodePointerFromFile() function in our "lfs" package to check whether
they should replace a file in the working tree with the contents of
a Git LFS object.  This function checks whether a file exists at a
specific location, and if it does, tries to parse the file's content as
a Git LFS pointer.

If no file is found, the object's data will be written to a new file
when the command calls the SmudgeToFile() method of the GitFilter
structure in the same "lfs" package.  Otherwise, the existing file will
only be overwritten with the object's data if the current file contents
are a valid Git LFS pointer to the given object.  Under other conditions,
such as when files have altered content or have been replaced by
directories, the commands should make no changes to the working tree.

The "checkout" and "pull" tests in our t/t-checkout.sh and t/t-pull.sh
test scripts, respectively, confirm that the commands will write new
files with the expected content when there are no files at the
corresponding locations in the working tree, and the "pull" test
also checks that no changes are made when a file has been replaced
with an empty one.

However, we do not have any tests which more explicitly test that
our "git lfs checkout" and "git lfs pull" commands do not alter files
that have been modified or replaced in the working tree.

Before we adjust how these commands check and create files, we add a
pair of new tests which validate that the commands make no changes when
a file has new content or has been replaced with a directory, including
when the commands are run from within such a directory.  Our tests also
use the assert_clean_index() test helper function we introduced in a
previous commit to confirm that the commands do not add the paths of
the changed files to the Git index.

By changing the current working directory to be a directory that has
replaced a Git LFS file, our tests verify that the "git lfs checkout"
and "git lfs pull" commands will not remove the directory within which
the user ran the commands.  While this is not especially significant
at the moment, it will become more so in a subsequent commit when
we revise our commands to change their current working directory to
the root of the work tree before writing any files.