t/t-{checkout,pull}.sh: read-only test for pull
In commit 686bda3722f12293f345240532f666b6a0961bb2 of PR #3120 we
revised the SmudgeToFile() method of the GitFilter structure in our
"lfs" package to try to update to read-only files while preserving
their read-only permissions.

The SmudgeToFile() method was changed so it starts by calling the
Stat() function of the Go standard library's "os" package to read
the file's current permissions, and then calls the Chmod() function
to add write permissions if the file does not have them yet.  If that
succeeds, a deferred call to the Chmod() function is registered which
will remove write permissions from the file when our method returns.

This change was made specifically to handle the case which may occur
when our "lockable" Git attribute applies to a given path, the file
at that path contains only a Git LFS pointer and not the corresponding
object data, and the user has not used the "git lfs lock" command to
acquire a lock on the file, so the file has read-only permissions.

In the same commit, we also added a new test to what is now our
t/t-checkout.sh test script to validate the new behaviour of the
"git lfs checkout" command.  However, we did not add an equivalent
test of the "git lfs pull" command to its test script, although this
command should exhibit the same behaviour when handling read-only
files as the "git lfs checkout" command.

As we expect to alter how the SmudgeToFile() method and its callers
check and create files in subsequent commits, we first add a test
to the t/t-pull.sh test script which verifies the "git lfs pull"
command's treatment of read-only files.

We also update the existing test in the t/t-checkout.sh test script
to correct a typo in the test's name, to add an "a" symbol to the
chmod(1) command's argument to clarify that it removes write permissions
for all users, and to use the same naming scheme as a number of our
other tests for the directory into which the test repository is cloned.