t/t-checkout.sh: add empty file mtime test
In commit 4c782ba772ed0137a69bd3a61b74c2cfb3805fd2 of PR #5491 we
revised the SmudgeToFile() method of the GitFilter structure in our
"lfs" package to try to avoid altering the modification timestamps
of empty files when they correspond to empty files in the Git reference
being checked out by the "git lfs pull" command.

The SmudgeToFile() method was changed so that if the Stat() function
from the Go standard library finds a file with a size of zero, and the
Git LFS pointer passed to the SmudgeToFile() method also has a size of
zero, the method takes no further action and simply returns.

In the same commit we also added a new test to the t/t-pull.sh test
script which validates the new behaviour of the "git lfs pull" command.
However, we did not add an equivalent test of the "git lfs checkout"
command, although this command should exhibit the same behaviour when
handling empty files as the "git lfs pull" command.

As we expect to alter how the SmudgeToFile() method and its callers
check and create files in subsequent commits, we first add a test to
the t/t-checkout.sh test script which verifies the "git lfs checkout"
command's treatment of empty files.