t: add checks of filepath filter cache option
In previous commits in this PR we implemented a LRU cache of file paths
that have matched against the set of file path filter patterns specified
for the current command, which may be any of our commands that accept
--include or --exclude options (or their -I and -X short forms) or read
the "lfs.fetchInclude" and "lfs.fetchExclude" configuration options.

We also implemented a new "lfs.pathFilterCacheSize" configuration option
which may be used to resize the cache as desired, or disable it entirely.

To validate whether the determineFilepathFilterCache() function we
introduced to our "commands" package is able to parse the range of
different values we accept for the "lfs.pathFilterCacheSize" option,
as well as handle invalid options, we update our shell test suite with
a new test in the t/t-ls-files.sh test script and a revised version
of another test in the t/t-migrate-import.sh test script.

These tests run the respective "git lfs ls-files" and "git lfs migrate
import" commands with a number of different settings of the
"lfs.pathFilterCacheSize" configuration option and confirm that the
same results are produced each time.