t/t-credentials.sh: also test .netrc on Windows
In previous commits in this PR we revised the ParseNetrc() function
in our "creds" package to support both ".netrc" and "_netrc" files
on Windows, whereas it has previously only supported "_netrc" files.

This change will align the behaviour of the Git LFS client with that
of contemporary versions of curl and Git, as they now support both
filenames on Windows.  Specifically, this support was added in commit
curl/curl@f9c7ba9096ec29db2536481d8e9ebe314e007f0c, which was included
in curl v7.66.0.

To validate these changes, we also update our t/t-credentials.sh test
script so it runs the key "credentials from netrc" test twice on Windows,
once with a ".netrc" file and once with a legacy "_netrc" file.

As well, we adjust the other tests which previously used a "_netrc"
file on Windows to simply use a ".netrc" file, as this should work
on all systems now.