config/netrc.go: drop unused extra netrc parser
Support for reading and parsing ".netrc" files (or "_netrc" files on
Windows) was first added to the Git LFS client in PR #715, and the
parsing code was later moved into our "config" package in PR #1226.

Subsequently, almost identical parsing code was also added to our
"lfsapi" package in commit 2275188dcb2e018da8d6dcde2205f6b539d39b41
of PR #1839, which was later moved to its current location in our
"creds" package in PR #3307.

Meanwhile, the only caller of the original ".netrc" parser, which
was implemented in the parseNetrc() method of the Configuration structure
in our "config" package, was removed in commit
afcd2118a44ede5d62a04283f0362727d45be18b PR #1846.

As a consequence, since PR #1846 the parseNetrc() method in the "config"
package has remained unused, along with the related "netrcfinder"
interface and associated structure and method in the same source file,
so we can simply delete that file now.