git: simplify UpdateRefsFromStdin() name
As suggested by larsxschneider on PR review, we can simplify the name
of the new UpdateRefsFromStdinInDir() function we added in a previous
commit in this PR to just UpdateRefsFromStdin(), without much loss
of clarity.

There is one other Git command wrapper function in our "git" package,
the AllRefsIn() function, which also sets the Git command's working
directory to an explicit value, which is why this function's name
has an "In" suffix, as did the UpdateRefsIn() function we removed
in a previous commit in this PR.

On the other hand, the NewLsFiles() and NewRevListScanner() constructor
functions in our "git" package also set the working directory for
the Git commands they perform, and their names do not explicitly
indicate this behaviour, so we have a precedent for just naming our
new function without any special suffix.  Like these constructor
functions, we rely on the function's argument list and the comments
which precede the function to clarify its actions.