script/lib/distro.rb: drop Debian 10 platform
As announced with the v3.6.0 release of the Git LFS client, and as
noted in commit 23fbfe8725553ad58ab8a50dad6fa19c6a602dc7 of PR #5911,
since all the distribution versions based on Debian 10 ("buster")
for which we previously built Debian packages have now reached the
end of their standard LTS (Long-Term Support) periods, future releases
of Git LFS will no longer provide support for this platform or build
packages for it.

Because we anticipate making a v3.7.0 release of the Git LFS client in
the near future, we removed the Dockerfile we used to build packages
for the Debian 10 platform in commit
git-lfs/build-dockers@04829f99e496ddf65388bb4468506ae049e86ab3 of
PR git-lfs/build-dockers#71.

We can now also remove the entries for this platform from the list
of our supported Linux distribution versions in the DistroMap Ruby
class source file, which is utilized by several scripts run by our
GitHub Actions release workflow jobs when building and naming the
Linux packages we publish on Packagecloud.